\name{f_1D}
\alias{f_1D}
\docType{data}
\title{
Output values of the evaluation of a function with one input variable and defined as a linear combination of B-splines
}
\description{
The evaluations were obtained by using the \eqn{\texttt{fda}} package with 201 input values contained in the xpred_1D.rda file.
}
\usage{data("f_1D")}
\format{
  Numeric vector of 201 values.
}