\name{f_2D}
\alias{f_2D}
\docType{data}
\title{
Output values of the evaluation of a function with two input variables and defined as a linear combination of a tensor product of univariate B-splines
}
\description{
The evaluations were obtained by using the \eqn{\texttt{fda}} package with 10000 input values contained in the xpred_2D.rda file.
}
\usage{data("f_2D")}
\format{
  Numeric vector of 10000 values.
}
