\name{glober-package}
\alias{glober}
\docType{package}
\title{
\packageTitle{glober}
}
\description{
glober consists of two functions: "glober.1d.R" and "glober.2d.R".
For further information on how to use these functions, 
we refer the reader to the vignette of the package.
}
\details{
Eight datasets are also provided within this package and used in examples of this manual and in the vignette. 
}
\author{
Mary E. Savino

Maintainer: Mary E. Savino <mary.savino@outlook.fr>
}
\references{
Savino, M. E. and Levy-Leduc, C. A novel approach for estimating functions in the multivariate setting based on an adaptive knot selection for B-splines with an application to a chemical system used in geoscience (2023), arXiv:2306.00686.
}
\keyword{package}

