\name{xpred_1D}
\alias{xpred_1D}
\docType{data}
\title{
Values of the single input variable for which a function has to be estimated
}
\description{
An example of input values containing the observation set and values for which an estimation of a given function \eqn{f} of one input variable is needed. 
}
\usage{data("xpred_1D")}
\format{
  Numeric vector of 201 values.
}