\name{xpred_2D}
\alias{xpred_2D}
\docType{data}
\title{
Values of the two input variables for which a function has to be estimated
}
\description{
An example of input values containing the observation set and values for which an estimation of a given function \eqn{f} of two input variables is needed. 
}
\usage{data("xpred_2D")}
\format{
Numeric matrix of 10000 rows and 2 columns.
}