\name{y_1D}
\alias{y_1D}
\docType{data}
\title{
Values of the response variable of the noisy observation set of one input variable
}
\description{
An example of noisy observations obtained by adding a Gaussian noise to \eqn{f(x_i)} contained in f_1D.rda and associated to the input values contained in x_1D.rda.
}
\usage{data("y_1D")}
\format{
  Numeric vector of 70 values.
}
