\name{y_2D}
\alias{y_2D}
\docType{data}
\title{
Values of the response variable of the noisy observation set of two input variables
}
\description{
An example of noisy observations obtained by adding a Gaussian noise to \eqn{f(x_i)} contained in f_2D.rda and associated to the input values contained in x_2D.rda.}
\usage{data("y_2D")}
\format{
  Numeric vector of 100 values.
}
